refactor: split industrial logistics modules
