+++ crates/oasis7/src/viewer/runtime_live/mapping.rs
@@ -19,6 +19,10 @@ use crate::simulator::{
 };
 use std::collections::BTreeMap;
 
+#[path = "mapping_helpers.rs"]
+mod mapping_helpers;
+pub(super) use mapping_helpers::runtime_event_kind_label;
+
 pub(super) fn runtime_state_to_simulator_model(
     state: &crate::runtime::WorldState,
     sidecar: &RuntimeLlmSidecar,
@@ -643,27 +647,6 @@ fn action_accepted_player_feedback(notes: &[String]) -> &'static str {
     } else {
         "Action accepted."
     }
-}
-
-fn runtime_event_kind_label(body: &RuntimeWorldEventBody) -> (String, Option<String>) {
-    let label = match body {
-        RuntimeWorldEventBody::Domain(_) => "domain",
-        RuntimeWorldEventBody::EffectQueued(_) => "effect_queued",
-        RuntimeWorldEventBody::ReceiptAppended(_) => "receipt_appended",
-        RuntimeWorldEventBody::PolicyDecisionRecorded(_) => "policy_decision_recorded",
-        RuntimeWorldEventBody::RuleDecisionRecorded(_) => "rule_decision_recorded",
-        RuntimeWorldEventBody::ActionOverridden(_) => "action_overridden",
-        RuntimeWorldEventBody::Governance(_) => "governance",
-        RuntimeWorldEventBody::ModuleEvent(_) => "module_event",
-        RuntimeWorldEventBody::ModuleCallFailed(_) => "module_call_failed",
-        RuntimeWorldEventBody::ModuleEmitted(_) => "module_emitted",
-        RuntimeWorldEventBody::ModuleStateUpdated(_) => "module_state_updated",
-        RuntimeWorldEventBody::ModuleRuntimeCharged(_) => "module_runtime_charged",
-        RuntimeWorldEventBody::SnapshotCreated(_) => "snapshot_created",
-        RuntimeWorldEventBody::ManifestUpdated(_) => "manifest_updated",
-        RuntimeWorldEventBody::RollbackApplied(_) => "rollback_applied",
-    };
-    (label.to_string(), None)
 }
 
 #[cfg(test)]
